+++ app/_definitions/models/entity-watchers/mom/MomInventoryOperation.ts
@@ -1,4 +1,10 @@
-import type {EntityWatcher, EntityWatchHandlerContext, IRpdServer} from "@ruiapp/rapid-core";
+import type {
+  ActionHandlerContext,
+  EntityWatcher,
+  EntityWatchHandlerContext,
+  IRpdServer,
+  RouteContext
+} from "@ruiapp/rapid-core";
 import {MomInventoryOperationType} from "~/_definitions/meta/data-dictionary-types";
 import {
   BaseLocation,
@@ -720,15 +726,17 @@ export default [
         }
 
         if (changes) {
-          await server.getEntityManager("sys_audit_log").createEntity({
-            entity: {
-              user: { id: ctx?.routerContext?.state.userId },
-              targetSingularCode: "mom_inventory_operation",
-              targetSingularName: `库存操作单 - ${ inventoryOperation?.businessType?.name } - ${ inventoryOperation?.code }`,
-              method: "update",
-              changes: changes,
-            }
-          })
+          if (ctx?.routerContext?.state.userId) {
+            await server.getEntityManager("sys_audit_log").createEntity({
+              entity: {
+                user: { id: ctx?.routerContext?.state.userId },
+                targetSingularCode: "mom_inventory_operation",
+                targetSingularName: `库存操作单 - ${ inventoryOperation?.businessType?.name } - ${ inventoryOperation?.code }`,
+                method: "update",
+                changes: changes,
+              }
+            })
+          }
         }
       } catch (e) {
         console.log(e)
@@ -754,14 +762,17 @@ export default [
           properties: ["id", "code", "businessType"],
         });
 
-        await server.getEntityManager("sys_audit_log").createEntity({
-          entity: {
-            user: { id: ctx?.routerContext?.state.userId },
-            targetSingularCode: "mom_inventory_operation",
-            targetSingularName: `库存操作单 - ${ inventoryOperation?.businessType?.name } - ${ inventoryOperation?.code }`,
-            method: "delete",
-          }
-        })
+        if (ctx?.routerContext?.state.userId) {
+          await server.getEntityManager("sys_audit_log").createEntity({
+            entity: {
+              user: { id: ctx?.routerContext?.state.userId },
+              targetSingularCode: "mom_inventory_operation",
+              targetSingularName: `库存操作单 - ${ inventoryOperation?.businessType?.name } - ${ inventoryOperation?.code }`,
+              method: "delete",
+            }
+          })
+        }
+
       } catch (e) {
         console.error(e);
       }
